+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Practices.Syncfusion.Excel` Update Syncfusion.XlsIO.Net.Core to 31.2.18
 - `Practices.Syncfusion.Pdf` Update Syncfusion.Pdf.Net.Core to 31.2.18
 - `practices-ui-ktbe` Added `whitespace-pre-line` CSS class to `PuibeReadonlyLabelValueComponent` to properly display line breaks in readonly values.
+- `practices-ng-forms` Updated `toFormControlOptions` nonNullable behavior.
+- _BREAKING_ `practices-ng-forms` Removed `AbstractControl` from `FormGroupDefinitionRecord` to limit configuration to supported properties.
 
 ### Fixed