+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `practices-ng-forms` Fixed form group fluent builder to apply `validators` and `asyncValidators` from control definitions.
+- `practices-ui-ktbe` Fixed `PuibeHideIfEmptyTextDirective` to work with projected elements in the directive tree.
 
 ## [11.1.0](https://github.com/nexplore/nexplore-practices/releases/tag/11.1.0) - 2025-12-10