+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [UNRELEASED]
+## [11.2.0](https://github.com/nexplore/nexplore-practices/releases/tag/11.2.0) - 2026-04-13
 
 ### Added